Extract search purging into new listener
