+++ core-bundle/tests/EventListener/DataContainer/PageUrlListenerTest.php
@@ -19,7 +19,6 @@ use Contao\CoreBundle\Tests\TestCase;
 use Contao\DataContainer;
 use Contao\Input;
 use Contao\PageModel;
-use Contao\Search;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Result;
 use PHPUnit\Framework\MockObject\MockObject;
@@ -748,142 +747,6 @@ class PageUrlListenerTest extends TestCase
         $this->expectExceptionMessage('Numeric aliases are not supported!');
 
         $listener->generateAlias('123', $dc);
-    }
-
-    public function testPurgesTheSearchIndexOnAliasChange(): void
-    {
-        $connection = $this->createMock(Connection::class);
-        $connection
-            ->expects($this->once())
-            ->method('fetchFirstColumn')
-            ->with('SELECT url FROM tl_search WHERE pid=:pageId', ['pageId' => 17])
-            ->willReturn(['uri'])
-        ;
-
-        $search = $this->mockAdapter(['removeEntry']);
-        $search
-            ->expects($this->once())
-            ->method('removeEntry')
-            ->with('uri')
-        ;
-
-        $dc = $this->mockClassWithProperties(
-            DataContainer::class,
-            [
-                'id' => 17,
-                'activeRecord' => (object) [
-                    'alias' => 'foo',
-                ],
-            ]
-        );
-
-        $listener = new PageUrlListener(
-            $this->mockContaoFramework([Search::class => $search]),
-            $this->createMock(Slug::class),
-            $this->createMock(TranslatorInterface::class),
-            $connection
-        );
-
-        $listener->purgeSearchIndexOnAliasChange('bar', $dc);
-    }
-
-    public function testDoesNotPurgeTheSearchIndexWithUnchangedAlias(): void
-    {
-        $connection = $this->createMock(Connection::class);
-        $connection
-            ->expects($this->never())
-            ->method($this->anything())
-        ;
-
-        $search = $this->mockAdapter(['removeEntry']);
-        $search
-            ->expects($this->never())
-            ->method($this->anything())
-        ;
-
-        $dc = $this->mockClassWithProperties(
-            DataContainer::class,
-            [
-                'id' => 17,
-                'activeRecord' => (object) [
-                    'alias' => 'foo',
-                ],
-            ]
-        );
-
-        $listener = new PageUrlListener(
-            $this->mockContaoFramework([Search::class => $search]),
-            $this->createMock(Slug::class),
-            $this->createMock(TranslatorInterface::class),
-            $connection
-        );
-
-        $listener->purgeSearchIndexOnAliasChange('foo', $dc);
-    }
-
-    public function testPurgesTheSearchIndexOnDelete(): void
-    {
-        $connection = $this->createMock(Connection::class);
-        $connection
-            ->expects($this->once())
-            ->method('fetchFirstColumn')
-            ->with('SELECT url FROM tl_search WHERE pid=:pageId', ['pageId' => 17])
-            ->willReturn(['uri'])
-        ;
-
-        $search = $this->mockAdapter(['removeEntry']);
-        $search
-            ->expects($this->once())
-            ->method('removeEntry')
-            ->with('uri')
-        ;
-
-        $dc = $this->mockClassWithProperties(
-            DataContainer::class,
-            [
-                'id' => 17,
-            ]
-        );
-
-        $listener = new PageUrlListener(
-            $this->mockContaoFramework([Search::class => $search]),
-            $this->createMock(Slug::class),
-            $this->createMock(TranslatorInterface::class),
-            $connection
-        );
-
-        $listener->purgeSearchIndexOnDelete($dc);
-    }
-
-    public function testDoesNotPurgeTheSearchIndexWithoutId(): void
-    {
-        $connection = $this->createMock(Connection::class);
-        $connection
-            ->expects($this->never())
-            ->method($this->anything())
-        ;
-
-        $search = $this->mockAdapter(['removeEntry']);
-        $search
-            ->expects($this->never())
-            ->method($this->anything())
-        ;
-
-        $dc = $this->mockClassWithProperties(
-            DataContainer::class,
-            [
-                'id' => null,
-            ]
-        );
-
-        $listener = new PageUrlListener(
-            $this->mockContaoFramework([Search::class => $search]),
-            $this->createMock(Slug::class),
-            $this->createMock(TranslatorInterface::class),
-            $connection
-        );
-
-        $listener->purgeSearchIndexOnDelete($dc);
     }
 
     public function testResetsThePrefixesAndSuffixes(): void
